dstp_backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
drf-spectacular==0.29.0
django-filter==25.2
django-extensions==4.1
mysqlclient==2.2.8
python-decouple==3.8
argon2-cffi==25.1.0
Pillow==12.1.1
